+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/sms/listener/TrackedEntityRegistrationSMSListener.java
@@ -109,7 +109,7 @@ public class TrackedEntityRegistrationSMSListener
                 patientAttributeValues.add( trackedEntityAttributeValue );
             });
 
-        int trackedEntityInstanceId = 0;
+        long trackedEntityInstanceId = 0;
         if ( patientAttributeValues.size() > 0 )
         {
             trackedEntityInstanceId = trackedEntityInstanceService.createTrackedEntityInstance( trackedEntityInstance,
