+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/userkeyjsonvalue/DefaultUserKeyJsonValueService.java
@@ -59,7 +59,7 @@ public class DefaultUserKeyJsonValueService
     }
 
     @Override
-    public int addUserKeyJsonValue( UserKeyJsonValue userKeyJsonValue )
+    public long addUserKeyJsonValue( UserKeyJsonValue userKeyJsonValue )
     {
         userKeyJsonValueStore.save( userKeyJsonValue );
         return userKeyJsonValue.getId();
