+++ dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/eventchart/EventChartServiceTest.java
@@ -71,9 +71,9 @@ public class EventChartServiceTest
         ecC.setProgram( prA );
         ecC.setType( ChartType.COLUMN );
 
-        int idA = eventChartService.saveEventChart( ecA );
-        int idB = eventChartService.saveEventChart( ecB );
-        int idC = eventChartService.saveEventChart( ecC );
+        long idA = eventChartService.saveEventChart( ecA );
+        long idB = eventChartService.saveEventChart( ecB );
+        long idC = eventChartService.saveEventChart( ecC );
 
         assertEquals( "ecA", eventChartService.getEventChart( idA ).getName() );
         assertEquals( "ecB", eventChartService.getEventChart( idB ).getName() );
@@ -93,9 +93,9 @@ public class EventChartServiceTest
         ecC.setProgram( prA );
         ecC.setType( ChartType.COLUMN );
 
-        int idA = eventChartService.saveEventChart( ecA );
-        int idB = eventChartService.saveEventChart( ecB );
-        int idC = eventChartService.saveEventChart( ecC );
+        long idA = eventChartService.saveEventChart( ecA );
+        long idB = eventChartService.saveEventChart( ecB );
+        long idC = eventChartService.saveEventChart( ecC );
 
         assertNotNull( eventChartService.getEventChart( idA ) );
         assertNotNull( eventChartService.getEventChart( idB ) );
